+++ proto/ibc/core/channel/v1/channel.proto
@@ -6,6 +6,39 @@ option go_package = "github.com/cosmos/cosmos-sdk/x/ibc/core/04-channel/types";
 import "gogoproto/gogo.proto";
 import "ibc/core/client/v1/client.proto";
 
+// Msg defines the ibc/channel Msg service.
+service Msg {
+  // ChannelOpenInit defines a rpc handler method for MsgChannelOpenInit.
+  rpc ChannelOpenInit(MsgChannelOpenInit) returns (MsgChannelOpenInitResponse);
+
+  // ChannelOpenTry defines a rpc handler method for MsgChannelOpenTry.
+  rpc ChannelOpenTry(MsgChannelOpenTry) returns (MsgChannelOpenTryResponse);
+
+  // ChannelOpenAck defines a rpc handler method for MsgChannelOpenAck.
+  rpc ChannelOpenAck(MsgChannelOpenAck) returns (MsgChannelOpenAckResponse);
+
+  // ChannelOpenConfirm defines a rpc handler method for MsgChannelOpenConfirm.
+  rpc ChannelOpenConfirm(MsgChannelOpenConfirm) returns (MsgChannelOpenConfirmResponse);
+
+  // ChannelCloseInit defines a rpc handler method for MsgChannelCloseInit.
+  rpc ChannelCloseInit(MsgChannelCloseInit) returns (MsgChannelCloseInitResponse);
+
+  // ChannelCloseConfirm defines a rpc handler method for MsgChannelCloseConfirm.
+  rpc ChannelCloseConfirm(MsgChannelCloseConfirm) returns (MsgChannelCloseConfirmResponse);
+
+  // RecvPacket defines a rpc handler method for MsgRecvPacket.
+  rpc RecvPacket(MsgRecvPacket) returns (MsgRecvPacketResponse);
+
+  // Timeout defines a rpc handler method for MsgTimeout.
+  rpc Timeout(MsgTimeout) returns (MsgTimeoutResponse);
+
+  // TimeoutOnClose defines a rpc handler method for MsgTimeoutOnClose.
+  rpc TimeoutOnClose(MsgTimeoutOnClose) returns (MsgTimeoutOnCloseResponse);
+
+  // Acknowledgement defines a rpc handler method for MsgAcknowledgement.
+  rpc Acknowledgement(MsgAcknowledgement) returns (MsgAcknowledgementResponse);
+}
+
 // MsgChannelOpenInit defines an sdk.Msg to initialize a channel handshake. It
 // is called by a relayer on Chain A.
 message MsgChannelOpenInit {
@@ -18,7 +51,10 @@ message MsgChannelOpenInit {
   string  signer     = 4;
 }
 
-// MsgChannelOpenInit  defines a msg sent by a Relayer to try to open a channel
+// MsgChannelOpenInitResponse defines the Msg/ChannelOpenInit response type.
+message MsgChannelOpenInitResponse {}
+
+// MsgChannelOpenInit defines a msg sent by a Relayer to try to open a channel
 // on Chain B.
 message MsgChannelOpenTry {
   option (gogoproto.equal)           = false;
@@ -34,6 +70,9 @@ message MsgChannelOpenTry {
       [(gogoproto.moretags) = "yaml:\"proof_height\"", (gogoproto.nullable) = false];
   string signer = 8;
 }
+
+// MsgChannelOpenTryResponse defines the Msg/ChannelOpenTry response type.
+message MsgChannelOpenTryResponse {}
 
 // MsgChannelOpenAck defines a msg sent by a Relayer to Chain A to acknowledge
 // the change of channel state to TRYOPEN on Chain B.
@@ -51,6 +90,9 @@ message MsgChannelOpenAck {
   string signer = 7;
 }
 
+// MsgChannelOpenAckResponse defines the Msg/ChannelOpenAck response type.
+message MsgChannelOpenAckResponse {}
+
 // MsgChannelOpenConfirm defines a msg sent by a Relayer to Chain B to
 // acknowledge the change of channel state to OPEN on Chain A.
 message MsgChannelOpenConfirm {
@@ -65,6 +107,9 @@ message MsgChannelOpenConfirm {
   string signer = 5;
 }
 
+// MsgChannelOpenConfirmResponse defines the Msg/ChannelOpenConfirm response type.
+message MsgChannelOpenConfirmResponse {}
+
 // MsgChannelCloseInit defines a msg sent by a Relayer to Chain A
 // to close a channel with Chain B.
 message MsgChannelCloseInit {
@@ -75,6 +120,9 @@ message MsgChannelCloseInit {
   string channel_id = 2 [(gogoproto.moretags) = "yaml:\"channel_id\""];
   string signer     = 3;
 }
+
+// MsgChannelCloseInitResponse defines the Msg/ChannelCloseInit response type.
+message MsgChannelCloseInitResponse {}
 
 // MsgChannelCloseConfirm defines a msg sent by a Relayer to Chain B
 // to acknowledge the change of channel state to CLOSED on Chain A.
@@ -90,6 +138,9 @@ message MsgChannelCloseConfirm {
   string signer = 5;
 }
 
+// MsgChannelCloseConfirmResponse defines the Msg/ChannelCloseConfirm response type.
+message MsgChannelCloseConfirmResponse {}
+
 // MsgRecvPacket receives incoming IBC packet
 message MsgRecvPacket {
   option (gogoproto.equal)           = false;
@@ -101,6 +152,9 @@ message MsgRecvPacket {
       [(gogoproto.moretags) = "yaml:\"proof_height\"", (gogoproto.nullable) = false];
   string signer = 4;
 }
+
+// MsgRecvPacketResponse defines the Msg/RecvPacket response type.
+message MsgRecvPacketResponse {}
 
 // MsgTimeout receives timed-out packet
 message MsgTimeout {
@@ -114,6 +168,9 @@ message MsgTimeout {
   uint64 next_sequence_recv = 4 [(gogoproto.moretags) = "yaml:\"next_sequence_recv\""];
   string signer             = 5;
 }
+
+// MsgTimeoutResponse defines the Msg/Timeout response type.
+message MsgTimeoutResponse {}
 
 // MsgTimeoutOnClose timed-out packet upon counterparty channel closure.
 message MsgTimeoutOnClose {
@@ -129,6 +186,9 @@ message MsgTimeoutOnClose {
   string signer             = 6;
 }
 
+// MsgTimeoutOnCloseResponse defines the Msg/TimeoutOnClose response type.
+message MsgTimeoutOnCloseResponse {}
+
 // MsgAcknowledgement receives incoming IBC acknowledgement
 message MsgAcknowledgement {
   option (gogoproto.equal)           = false;
@@ -141,6 +201,9 @@ message MsgAcknowledgement {
       [(gogoproto.moretags) = "yaml:\"proof_height\"", (gogoproto.nullable) = false];
   string signer = 5;
 }
+
+// MsgAcknowledgementResponse defines the Msg/Acknowledgement response type.
+message MsgAcknowledgementResponse {}
 
 // Channel defines pipeline for exactly-once packet delivery between specific
 // modules on separate blockchains, which has at least one end capable of
