+++ proto/ibc/core/connection/v1/connection.proto
@@ -8,6 +8,21 @@ import "google/protobuf/any.proto";
 import "ibc/core/commitment/v1/commitment.proto";
 import "ibc/core/client/v1/client.proto";
 
+// Msg defines the ibc/connection Msg service.
+service Msg {
+  // ConnectionOpenInit defines a rpc handler method for MsgConnectionOpenInit.
+  rpc ConnectionOpenInit(MsgConnectionOpenInit) returns (MsgConnectionOpenInitResponse);
+
+  // ConnectionOpenTry defines a rpc handler method for MsgConnectionOpenTry.
+  rpc ConnectionOpenTry(MsgConnectionOpenTry) returns (MsgConnectionOpenTryResponse);
+
+  // ConnectionOpenAck defines a rpc handler method for MsgConnectionOpenAck.
+  rpc ConnectionOpenAck(MsgConnectionOpenAck) returns (MsgConnectionOpenAckResponse);
+  
+  // ConnectionOpenConfirm defines a rpc handler method for MsgConnectionOpenConfirm.
+  rpc ConnectionOpenConfirm(MsgConnectionOpenConfirm) returns (MsgConnectionOpenConfirmResponse);
+}
+
 // MsgConnectionOpenInit defines the msg sent by an account on Chain A to
 // initialize a connection with Chain B.
 message MsgConnectionOpenInit {
@@ -20,6 +35,9 @@ message MsgConnectionOpenInit {
   string       version       = 4;
   string       signer        = 5;
 }
+
+// MsgConnectionOpenInitResponse defines the Msg/ConnectionOpenInit response type.
+message MsgConnectionOpenInitResponse { }
 
 // MsgConnectionOpenTry defines a msg sent by a Relayer to try to open a
 // connection on Chain B.
@@ -47,6 +65,9 @@ message MsgConnectionOpenTry {
   string signer = 12;
 }
 
+// MsgConnectionOpenTryResponse defines the Msg/ConnectionOpenTry response type.
+message MsgConnectionOpenTryResponse { }
+
 // MsgConnectionOpenAck defines a msg sent by a Relayer to Chain A to
 // acknowledge the change of connection state to TRYOPEN on Chain B.
 message MsgConnectionOpenAck {
@@ -71,6 +92,9 @@ message MsgConnectionOpenAck {
   string signer = 10;
 }
 
+// MsgConnectionOpenAckResponse defines the Msg/ConnectionOpenAck response type.
+message MsgConnectionOpenAckResponse { }
+
 // MsgConnectionOpenConfirm defines a msg sent by a Relayer to Chain B to
 // acknowledge the change of connection state to OPEN on Chain A.
 message MsgConnectionOpenConfirm {
@@ -84,6 +108,9 @@ message MsgConnectionOpenConfirm {
       [(gogoproto.moretags) = "yaml:\"proof_height\"", (gogoproto.nullable) = false];
   string signer = 4;
 }
+
+// MsgConnectionOpenConfirmResponse defines the Msg/ConnectionOpenConfirm response type.
+message MsgConnectionOpenConfirmResponse { }
 
 // ICS03 - Connection Data Structures as defined in
 // https://github.com/cosmos/ics/tree/master/spec/ics-003-connection-semantics#data-structures
